fixed requested test files and moved the App.test.js to the test folder, updated ReadMe
