+++ README.md
@@ -1,38 +1,24 @@
-# GameVerse
-
+# GameVerse FrontEnd
 ## OverView
 GameVerse is a gamified website that serves as a comprehensive catalog for games. The primary focus is on game reviews, allowing users to explore and discover new games while providing their own reviews and ratings.
-
-**Live Link** : [link to Github!](https://github.com/Yuhje-Gaming/gameverse-backend)
-
+**Live Link** : *Will add Render Link*
 # Features
-
 - User-Friendly Interface: GameVerse offers an intuitive and easy-to-use interface for seamless navigation.
-
 - Carousel of Games: The homepage showcases a dynamic carousel highlighting the hottest and trending games.
-
 - Catalog of Games on Index Page: Explore a comprehensive catalog of games right on the index page, covering everything from indie gems to AAA blockbusters.
-
 - Personalized Favorite Games: Users can curate their own collection by adding and removing favorite games, tailoring their experience to their preferences.
-
 - Game Management: Contribute to the platform's main database by creating, editing, and removing games, ensuring an up-to-date catalog.
-
 - Informative "About Us" Page: Learn about the backstory, mission, and passionate team behind GameVerse on an engaging "About Us" page.
-
-# Tech Used / Dependencies
-- "@testing-library/jest-dom": "^5.14.1",
-- "@testing-library/react": "^13.0.0",
-- "@testing-library/user-event": "^13.2.1",
-- "bootstrap": "^5.3.0",
-- "react": "^18.2.0",
-- "react-dom": "^18.2.0",
-- "react-router-dom": "^6.13.0",
-- "react-scripts": "5.0.1",
-- "reactstrap": "^9.2.0",
-- "web-vitals": "^2.1.0"
-
-# React Front End
-
-## Initial Configurations
-
-
+## Runnig the project locally
+**Live Version**
+[link to Github!](https://github.com/Yuhje-Gaming/gameverse-frontend)
+1.  Clone this project locally
+2.  Run yarn install
+3.  Host a catalog of games and invite some friends!!
+### ToDo List
+1.  Component and Page Testing
+2.  Component and Page Functionality
+### Contributors
+1.  Henri Lambertucci
+2.  Justin Karpman
+3.  Yue Han Lee
