+++ src/__tests__/Footer.test.js
@@ -1,0 +1,14 @@
+import Footer from "../components/Footer";
+import { render } from '@testing-library/react';
+import { BrowserRouter } from 'react-router-dom'
+
+describe('<Has a page that renders without crashing />', () => {
+  it('renders text for footer', () => {
+    render(
+      <BrowserRouter>
+        <Footer />
+      </BrowserRouter>
+    )
+  })
+})
+
